pyproject.toml
[project]
name = "pdf_segmentation"
version = "1.1.0"
description = ""
authors = [
    {name = "Luciano Bermudez",email = "lucianobmecheng69@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "langchain (>=1.2.8,<2.0.0)",
    "pymupdf (>=1.26.7,<2.0.0)",
    "langgraph (>=1.0.7,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "langchain-openai (>=1.1.7,<2.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)"
]

